+++ src/components/DarkModeToggleButton/darkModeToggleButton.module.scss
@@ -1,0 +1,15 @@
+.toggleThemeBtn {
+  display: flex;
+  justify-content: center;
+  width: 48px;
+  height: 48px;
+  cursor: pointer;
+  background-color: var(--color-heighlight);
+  border-radius: 100%;
+
+  &:hover {
+    background-color: var(--color-overlay);
+    transition: ease 500ms;
+    transform: translateY(-2px);
+  }
+}
