+++ src/components/SideBar/index.tsx
@@ -1,7 +1,4 @@
-'use client';
-
 import Image from 'next/image';
-import { useEffect, useState } from 'react';
 import SnsButtons from '../SnsButtons';
 import styles from './sidebar.module.scss';
 import Typography from '../Typography';
@@ -33,26 +30,8 @@ const MOCKDATA: IUerData = {
   ],
 };
 export default function SideBar() {
-  const [theme, setTheme] = useState<string>('light');
-
-  const onToggleTheme = () => {
-    const newTheme = theme === 'light' ? 'dark' : 'light';
-    setTheme(newTheme);
-  };
-
-  useEffect(() => {
-    document.documentElement.setAttribute('data-theme', theme);
-  }, [theme]);
-
   return (
     <aside className={styles.sidebarContainer}>
-      <button type="button" onClick={onToggleTheme} className={styles.toggleThemeBtn}>
-        {theme === 'light' ? (
-          <Image src="/svgs/dark.svg" alt="toggleButton" width={48} height={48} />
-        ) : (
-          <Image src="/svgs/light.svg" alt="toggeleButton" width={48} height={48} />
-        )}
-      </button>
       <Typography variant="h3" className={styles.sidebarTitle}>
         눈에 띄지 않는 것의 가치
       </Typography>
